% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anes_2020}
\alias{anes_2020}
\title{American National Election Studies (ANES) (2020) data}
\format{
A data frame with 7453 rows and 63 variables:
\describe{
\item{\code{V200001}}{double 2020 Case ID}
\item{\code{CaseID}}{double 2020 Case ID}
\item{\code{V200002}}{double Mode of interview: pre-election interview}
\item{\code{InterviewMode}}{factor Mode of interview: pre-election interview}
\item{\code{V200010b}}{double Full sample post-election weight}
\item{\code{Weight}}{double Full sample post-election weight}
\item{\code{V200010c}}{double Full sample variance unit}
\item{\code{VarUnit}}{factor Full sample variance unit}
\item{\code{V200010d}}{double Full sample variance stratum}
\item{\code{Stratum}}{factor Full sample variance stratum}
\item{\code{V201006}}{double PRE: How interested in following campaigns}
\item{\code{CampaignInterest}}{factor PRE: How interested in following campaigns}
\item{\code{V201024}}{double PRE: In what manner did R vote}
\item{\code{V201025x}}{double PRE: SUMMARY: Registration and early vote status}
\item{\code{V201029}}{double PRE: For whom did R vote for President}
\item{\code{V201101}}{double PRE: Did R vote for President in 2016 \link{revised}}
\item{\code{V201102}}{double PRE: Did R vote for President in 2016}
\item{\code{VotedPres2016}}{factor PRE: Did R vote for President in 2016}
\item{\code{V201103}}{double PRE: Recall of last (2016) Presidential vote choice}
\item{\code{VotedPres2016_selection}}{factor PRE: Recall of last (2016) Presidential vote choice}
\item{\code{V201228}}{double PRE: Party ID: Does R think of self as Democrat, Republican, or Independent}
\item{\code{V201229}}{double PRE: Party Identification strong - Democrat Republican}
\item{\code{V201230}}{double PRE: No Party Identification - closer to Democratic Party or Republican Party}
\item{\code{V201231x}}{double PRE: SUMMARY: Party ID}
\item{\code{PartyID}}{factor PRE: SUMMARY: Party ID}
\item{\code{V201233}}{double PRE: How often trust government in Washington to do what is right \link{revised}}
\item{\code{TrustGovernment}}{factor PRE: How often trust government in Washington to do what is right \link{revised}}
\item{\code{V201237}}{double PRE: How often can people be trusted}
\item{\code{TrustPeople}}{factor PRE: How often can people be trusted}
\item{\code{V201507x}}{double PRE: SUMMARY: Respondent age}
\item{\code{Age}}{double PRE: SUMMARY: Respondent age}
\item{\code{AgeGroup}}{factor PRE: SUMMARY: Respondent age (categorized)}
\item{\code{V201510}}{double PRE: Highest level of Education}
\item{\code{Education}}{factor PRE: Highest level of Education}
\item{\code{V201546}}{double PRE: R: Are you Spanish, Hispanic, or Latino}
\item{\code{V201547a}}{double RESTRICTED: PRE: Race of R: White \link{mention}}
\item{\code{V201547b}}{double RESTRICTED: PRE: Race of R: Black or African-American \link{mention}}
\item{\code{V201547c}}{double RESTRICTED: PRE: Race of R: Asian \link{mention}}
\item{\code{V201547d}}{double RESTRICTED: PRE: Race of R: Native Hawaiian or Pacific Islander \link{mention}}
\item{\code{V201547e}}{double RESTRICTED: PRE: Race of R: Native American or Alaska Native \link{mention}}
\item{\code{V201547z}}{double RESTRICTED: PRE: Race of R: other specify}
\item{\code{V201549x}}{double PRE: SUMMARY: R self-identified race/ethnicity}
\item{\code{RaceEth}}{factor PRE: SUMMARY: R self-identified race/ethnicity}
\item{\code{V201600}}{double PRE: What is your (R) sex? \link{revised}}
\item{\code{Gender}}{factor PRE: What is your (R) sex? \link{revised}}
\item{\code{V201607}}{double RESTRICTED: PRE: Total income amount - revised}
\item{\code{V201610}}{double RESTRICTED: PRE: Income amt missing - categories lt 20K}
\item{\code{V201611}}{double RESTRICTED: PRE: Income amt missing - categories 20-40K}
\item{\code{V201613}}{double RESTRICTED: PRE: Income amt missing - categories 40-70K}
\item{\code{V201615}}{double RESTRICTED: PRE: Income amt missing - categories 70-100K}
\item{\code{V201616}}{double RESTRICTED: PRE: Income amt missing - categories 100+K}
\item{\code{V201617x}}{double PRE: SUMMARY: Total (family) income}
\item{\code{Income}}{factor PRE: SUMMARY: Total (family) income (21 categories)}
\item{\code{Income7}}{factor PRE: SUMMARY: Total (family) income (7 categories)}
\item{\code{V202051}}{double POST: R registered to vote (post-election)}
\item{\code{V202066}}{double POST: Did R vote in November 2020 election}
\item{\code{V202072}}{double POST: Did R vote for President}
\item{\code{VotedPres2020}}{factor POST: Did R vote for President}
\item{\code{V202073}}{double POST: For whom did R vote for President}
\item{\code{V202109x}}{double PRE-POST: SUMMARY: Voter turnout in 2020}
\item{\code{V202110x}}{double PRE-POST: SUMMARY: 2020 Presidential vote}
\item{\code{VotedPres2020_selection}}{factor PRE-POST: SUMMARY: 2020 Presidential vote}
\item{\code{EarlyVote2020}}{factor Indicator of early voting in 2020}
}
}
\source{
\url{https://electionstudies.org/data-center/2020-time-series-study/}
}
\usage{
anes_2020
}
\description{
Based on the 2020 ANES data with derived variables and subset to people who completed both pre and post-election interviews
}
\keyword{datasets}
