% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anes_2024}
\alias{anes_2024}
\title{American National Election Studies (ANES) (2024) data}
\format{
A data frame with 4964 rows and 54 variables:
\describe{
\item{\code{V240001}}{double 2024 Case ID}
\item{\code{CaseID}}{double CASEID}
\item{\code{InterviewMode_Pre}}{Factor MODE OF INTERVIEW: PRE-ELECTION INTERVIEW}
\item{\code{InterviewMode_Post}}{Factor MODE OF INTERVIEW: POST-ELECTION INTERVIEW}
\item{\code{Weight}}{double FULL SAMPLE POST-ELECTION WEIGHT}
\item{\code{Stratum}}{double FULL SAMPLE VARIANCE STRATUM}
\item{\code{VarUnit}}{double FULL SAMPLE VARIANCE UNIT}
\item{\code{CampaignInterest}}{Factor PRE: HOW INTERESTED IN FOLLOWING CAMPAIGNS}
\item{\code{VotedPres2020}}{Factor PRE: DID R VOTE FOR PRESIDENT IN 2020 (REVISED)}
\item{\code{VotedPres2020_selection}}{Factor PRE: RECALL OF LAST (2020) PRESIDENTIAL VOTE CHOICE}
\item{\code{PartyID}}{Factor PRE: SUMMARY: PARTY ID}
\item{\code{TrustGovernment}}{Factor PRE: HOW OFTEN TRUST GOVERNMENT IN WASHINGTON TO DO WHAT IS RIGHT (REVISED)}
\item{\code{TrustPeople}}{Factor PRE: HOW OFTEN CAN PEOPLE BE TRUSTED}
\item{\code{Age}}{double PRE: SUMMARY: RESPONDENT AGE}
\item{\code{Education}}{double PRE: HIGHEST LEVEL OF EDUCATION}
\item{\code{RaceEth}}{Factor PRE: SUMMARY: R SELF-IDENTIFIED RACE/ETHNICITY}
\item{\code{Sex}}{Factor PRE: WHAT IS YOUR (R) SEX? (REVISED)}
\item{\code{Income}}{Factor PRE: SUMMARY: TOTAL (FAMILY) INCOME}
\item{\code{EarlyVote2024}}{Factor PRE: DID R ALREADY VOTE}
\item{\code{VotedPres2024}}{Factor POST: DID R VOTE FOR PRESIDENT}
\item{\code{VotedPres2024_selection}}{Factor PRE-POST: SUMMARY: 2024 PRESIDENTIAL VOTE}
\item{\code{AgeGroup}}{Factor PRE: SUMMARY: RESPONDENT AGE (categorized)}
\item{\code{EducationGroup}}{Factor PRE: HIGHEST LEVEL OF EDUCATION (categorized)}
\item{\code{Income7}}{Factor PRE: SUMMARY: TOTAL (FAMILY) INCOME (7 categories)}
\item{\code{V240001}}{double CASEID}
\item{\code{V240002a}}{double MODE OF INTERVIEW: PRE-ELECTION INTERVIEW}
\item{\code{V240002b}}{double MODE OF INTERVIEW: POST-ELECTION INTERVIEW}
\item{\code{V240108b}}{double FULL SAMPLE POST-ELECTION WEIGHT}
\item{\code{V240108d}}{double FULL SAMPLE VARIANCE STRATUM}
\item{\code{V240108c}}{double FULL SAMPLE VARIANCE UNIT}
\item{\code{V241005}}{double PRE: HOW INTERESTED IN FOLLOWING CAMPAIGNS}
\item{\code{V241103}}{double PRE: DID R VOTE FOR PRESIDENT IN 2020 (REVISED)}
\item{\code{V241104}}{double PRE: RECALL OF LAST (2020) PRESIDENTIAL VOTE CHOICE}
\item{\code{V241038}}{double PRE: DID R VOTE FOR PRESIDENT}
\item{\code{V241221}}{double PRE: PARTY ID}
\item{\code{V241222}}{double PRE: PARTY ID STRONG}
\item{\code{V241223}}{double PRE: PARTY ID LEAN}
\item{\code{V241227x}}{double PRE: SUMMARY: PARTY ID}
\item{\code{V241229}}{double PRE: HOW OFTEN TRUST GOVERNMENT IN WASHINGTON TO DO WHAT IS RIGHT (REVISED)}
\item{\code{V241234}}{double PRE: HOW OFTEN CAN PEOPLE BE TRUSTED}
\item{\code{V241458x}}{double PRE: SUMMARY: RESPONDENT AGE}
\item{\code{V241463}}{double PRE: HIGHEST LEVEL OF EDUCATION}
\item{\code{V241499}}{double PRE: HISPANIC ETHNICITY}
\item{\code{V241501x}}{double PRE: SUMMARY: R SELF-IDENTIFIED RACE/ETHNICITY}
\item{\code{V241550}}{double PRE: WHAT IS YOUR (R) SEX? (REVISED)}
\item{\code{V241566x}}{double PRE: SUMMARY: TOTAL (FAMILY) INCOME}
\item{\code{V242065}}{double POST: DID R VOTE IN NOVEMBER 2024 ELECTION}
\item{\code{V241035}}{double PRE: DID R ALREADY VOTE}
\item{\code{V241037}}{double PRE: HOW VOTE}
\item{\code{V242051}}{double POST: REGISTERED TO VOTE}
\item{\code{V242095x}}{double PRE-POST: SUMMARY: VOTER TURNOUT IN 2020}
\item{\code{V242066}}{double POST: DID R VOTE FOR PRESIDENT}
\item{\code{V241039}}{double PRE: FOR WHOM DID R VOTE FOR PRESIDENT (2024)}
\item{\code{V242067}}{double POST: FOR WHOM DID R VOTE FOR PRESIDENT (2024)}
\item{\code{V242096x}}{double PRE-POST: SUMMARY: 2024 PRESIDENTIAL VOTE}
}
}
\source{
\url{https://electionstudies.org/data-center/2024-time-series-study/}
}
\usage{
anes_2024
}
\description{
Based on the 2024 ANES data with derived variables and subset to people who completed both pre and post-election interviews
}
\keyword{datasets}
