% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncvs_2021_household}
\alias{ncvs_2021_household}
\title{National Crime Victimization Survey (NCVS) (2021) household file}
\format{
A data frame with 256460 rows and 12 variables:
\describe{
\item{\code{YEARQ}}{double YEAR AND QUARTER OF INTERVIEW (YYYY.Q)}
\item{\code{IDHH}}{character NCVS ID FOR HOUSEHOLDS}
\item{\code{WGTHHCY}}{double ADJUSTED HOUSEHOLD WEIGHT - COLLECTION YEAR}
\item{\code{V2117}}{double PSEUDOSTRATUM CODE}
\item{\code{V2118}}{double ECUCODE: HALF SAMPLE CODE}
\item{\code{V2015}}{integer TENURE (ALLOCATED)}
\item{\code{V2143}}{integer URBANICITY (START 2020 Q1)}
\item{\code{SC214A}}{integer HOUSEHOLD INCOME (ALLOCATED) (START 2015, Q1)}
\item{\code{V2122}}{integer FAMILY STRUCTURE CODE}
\item{\code{V2126B}}{integer PLACE SIZE CODE - 1990, 2000, 2010 SAMPLE DESIGN (START 1995 Q3)}
\item{\code{V2127B}}{integer REGION - 1990, 2000, 2010 SAMPLE DESIGN (START 1995 Q3)}
\item{\code{V2129}}{integer CBSA MSA STATUS}
}
}
\source{
\url{https://doi.org/10.3886/ICPSR38429.v1}
}
\usage{
ncvs_2021_household
}
\description{
A subset of the NCVS 2021 household file. Includes all records but a subset of columns
}
\keyword{datasets}
