% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncvs_2021_person}
\alias{ncvs_2021_person}
\title{National Crime Victimization Survey (NCVS) (2021) person file}
\format{
A data frame with 291878 rows and 11 variables:
\describe{
\item{\code{YEARQ}}{double YEAR AND QUARTER OF INTERVIEW (YYYY.Q)}
\item{\code{IDHH}}{character NCVS ID FOR HOUSEHOLDS}
\item{\code{IDPER}}{character NCVS ID FOR PERSONS}
\item{\code{WGTPERCY}}{double ADJUSTED PERSON WEIGHT - COLLECTION YEAR}
\item{\code{V3014}}{double AGE (ALLOCATED)}
\item{\code{V3015}}{integer MARITAL STATUS (CURRENT SURVEY)}
\item{\code{V3018}}{integer SEX (ALLOCATED)}
\item{\code{V3023A}}{integer RACE RECODE (START 2003 Q1)}
\item{\code{V3024}}{integer HISPANIC ORIGIN}
\item{\code{V3084}}{integer SEXUAL ORIENTATION (START 2017 Q1)}
\item{\code{V3086}}{integer CURRENT GENDER IDENTITY (START 2017 Q1)}
}
}
\source{
\url{https://doi.org/10.3886/ICPSR38429.v1}
}
\usage{
ncvs_2021_person
}
\description{
A subset of the NCVS 2021 person file. Includes all records but a subset of columns
}
\keyword{datasets}
