% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recs_2020_raw}
\alias{recs_2020_raw}
\title{Residential Energy Consumption Survey (RECS) (2020) data}
\format{
A data frame with 18496 rows and 789 variables. See "Variable and response codebook" at \url{https://www.eia.gov/consumption/residential/data/2020/index.php?view=microdata} for variable information.
}
\source{
\url{https://www.eia.gov/consumption/residential/data/2020/index.php?view=microdata}
}
\usage{
recs_2020_raw
}
\description{
2020 RECS microdata
}
\keyword{datasets}
